NLP/ACL2020-GraphSum/requirements.txt
nltk==3.4.5
numpy==1.18.1
paddlepaddle-gpu==1.6.3.post107
pyrouge==0.1.3
regex==2020.2.20
requests==2.22.0
sentencepiece==0.1.85